Generic witchcraft
We genericise ConnectionHandler and RequestHandler so that they just
handle dispatch to a regular function, a bit like the original
hyper::service::{make_service_fn, service_fn} did. However, we handle the
cloning of the channel from the websocket to the application in our impl.

Yes, the types look disgusting, because rust doesn't support trait aliases.

Oh, and the _reason_ I'm doing this is to try to separate out some of the
application concerns into different modules.